test/libcephfs: move the quota test to a dedicated file
